Add a check for hardware GL
Check the GL renderer string to see if it is "Software Rasterizer" to
avoid the user trying to switch to Compiz or GNOME Shell when it's
going to be vastly too slow.

Allow show_alert() to take secondary text in the GNOME style and use it
for the "Only the standard GNOME desktop is available" as well.